Reflect current tableschema
The consensus TableSchema model has changed to be in line with FrictionlessIO standards. The main change is lowercase column names, which breaks assumptions about column names made in the CfdeDataPackage. This commit changes the CfdeDataPackage to use the changed model, with lowercase column names.
In theory, these changes shouldn't break anything else.