Batch-load data into DERIVA
Large files will hang when POSTed to a DERIVA catalog, which is now mitigated by batching POST data.
One open question is what should the batch size be? Currently it's set at 50k and isn't configurable except by modifying the source code. This likely should change in the future, but is fine for now.