draft transitive-closure ETL tables for project and collection nesting
These provide a flattened leader-member association for each distinct
group of projects or collections which are "under" a given group
leader. They are reflexive, so the leader is a member of its own group
too.  These will allow easy join queries when you want "a project and
all its sub-projects" or "a collection and all its subset
collections" or the inverse "a project and all its super-projects"
or "a collection and all its superset collections".