add explicit INSERT INTO denorm_table_name (columns, ...) prefix to ETL queries
It seems expedient to include this domain-specific knowledge here in
the SQL files. It will be easier for consumers to trip the INSERT clause
if they want a standalone read-only query. To build the missing INSERT,
they would need to put a lot more effort into interpreting the SQL
themselves to infer the correct insertion clause.