first minimally working version of submission module
This represents a refactoring of prior submission pipeline to expose
utility layer and integrate with new submission registry.

1. Pre-validates invoking params
    - dcc_id known by registry
    - submitting_user has submitter role for dcc_id
2. Downloads bdbag from an "archive url"
    - Currently via anonymous HTTPS for test purposes
3. Registers submission
4. Unpacks and validates bdbag container
5. Does some CFDE-specific model checking
    - reports datapackage resources (tables) to registry
    - Expects a "subset" model compared to built-in portal model
6. Does frictionless.validate_package()
    - updates registry with per-table num_rows
7. Builds an empty review catalog
    - provisioned with built-in portal model + config
    - registers it w/ submission
8. Uploads content
    - uses existing CfdeDatapackage method
    - updates registry with per-table completion
9. Marks submission as content-ready in registry

Still pending is the extra work needed to prepare the derived tables
for the portal via ETL, as well as appropriate integration to work
with raw globus auth context as we expect to see in action provider
hosting environment.

The test harness in the submission module expects that deriva
credentials are available. This allows the sequence to be run from
the CLI as:

    python3 -m cfde_deriva.submission 'dcc_id' 'archive_url'

for testing, a fake submission UUID is derived via the hashed-URL
form of UUID, producing the same deterministic UUID for each
test run with the same input URL.