WIP snapshot: adding table_mask column to level1_stats
Extend the level1_stats ETL to produce the power set of results
for the possible subsets of the backbone query:

  file -- biosample -- subject
  file -- biosample
  biosample -- subject
  file
  biosample
  subject

Use an integer as a bit-field representing the set of core tables
used for a stats result.

Change the dashboard query helper to compute the corresponding mask
and set a table_mask=bits filter when querying level1_stats, to use
the most accurate rollup available for a given selection of stats cube
dimensions.

When completed, this will produce more accurate results including for
the simple cases like counting core entities w/o any vocabulary
grouping.