add support for datapackage.load_data_files(onconflict=update) and revise some vocab term titles in registry
To update a table (during reprovisioning), do two passes to first
insert missing rows, then to update non-key fields. Use this to apply
changes to presentation on built-in vocabulary etc.

Change titles of datapackage and approval status vocabulary terms to
make them a little more explicit in the web UI.