redefine synonyms as frictionless "array" (a JSON list)
This was previously defined as an odd list of term IDs and then
retracted from the C2M2 spec. The modeling group has discussed and
wants to return to a simpler human-readable synonyms set for UI term
discovery use cases.  Any structured relationships should be rendered
in the model as additional term-to-term associations.

Thus, for non-structured search, we want a simple list of
human-readable strings. It isn't clear that there is any frictionless
support to define per-value constraints on the array type.  In the
portal, we'll map this to a PostgreSQL text[] column type rather than
a general JSON type, to minimize what oddities we would store...