wip: sketch intermediate portal-prep ETL model
This assumes that the submission model will be augmented with serial nid
columns during load, and rewrites fkeys to be based on nid instead of
raw "id" or "id_namespace"+"local_id" values.

This also infers a mime_type table from the distinct file.mime_type
values in the submission.

This also introduces a core_fact table which conceptually replaces the
prior level1_stats with a set of truly disjoint facts representing
different combinations of vocabulary/project/namespace concepts.