add flags (w/ defaults) to make model validation more relaxed
When analyzing the datapackage model:

- tolerate non-nullability variations
- tolerate missing uniqueness constraints

This prevents an early exit, deferring checks until later when
the data itself must conform to the canonical submission model. This
approach allows a rolling upgrade where a datapackage may be able
to satisfy future requirements in practice, even if it doesn't
actually declare all the constraints properly.