clean up indexing_preferences hints to ermrest
This relies on a fix in latest ermrest, but means we can set a
datapackage-wide indexing_preference which will get mapped to an
annotation on the CFDE schema in the catalog, and ermrest will
use that as the default preferences for every table and column
which does not have a local override in its own definition.

This also assumes ermrest will produce btree indexes on every key
an foreign key, so we do not need to request those explicitly.