tweak indexing preferences with two-column indexes for likely joins
This relies on a new ermrest feature to accept a list of columns as
a hint for how to customize a btree index.

These provide a few "inverted" index structures which may help the
query planner for some search scenarios. For example, complementing
an association table's (fkey1, fkey2) uniqueness index with a
reverse (fkey2, fkey1) to walk subsets.