drop file/biosample/subject keywords columns
This will cause a temporary regression in the CI environment until
the associated Chaise enhancement is available.

Eventually, the file, biosample, and subject search pages will use the
keywords available via the references core_fact row, matching on all
the related CV and project info.

This dramatically reduces the storage overhead of on these tables
and associated indexes.