fixup "title" fields on multiple assocation tables
These aren't usually displayed in the portal, but if they are then we
want the actual display value for the "title" field, with proper
casing and whitespace.  The heuristics only work to convert a bare
model name into a display value, but will not be applied to a "title"
already set via annotation.