update submission and release card view and status info
- add "release status" to datapackage compact view
- apply similar card view to release compact view
- change display value for "planning" release status

the datapackage "release status" is a prioritized sort of the distinct
release status values reachable by ALL releases linked to the
datapackage.  the sort is based on the new release_status.dp_rank
ordinal column.  only the first status value in the sorted array
is displayed in the datapackage card view.