basic support for SQL-based data checks, file_check_checksums.sql as test
Generic support for an extra top-level field in the frictionless
resource (table) definition, sibling to "name" and "schema":

   "check_sql_paths": [ "check_file.sql" ]

in the built-in submission/c2m2-datapackage.json model file. The path
must be a basename in the same configs/submission directory.

The check will run _after_ basic frictionless validation has succeeded
and we have loaded the submission into SQLite, but _before_ our
foreign key checks which are also run against the SQLite DB.  The
referenced file MUST contain a single SQL statement which returns an
empty set for valid submissions or one error row for invalid
submissions.  The error row needs exactly 4 columns (understood
positionally):

1. a short human-readable description of the violated expectation
2. a count of the number of violating occurrences
3. a row number (nid) for the first failing row in this table
4. identifying data, such as a JSON object collecting relevant fields

The first implemented check is:

  configs/submission/file_check_checksums.sql

which enforces the rule that checksums are required for files
with a persistent_id value.

An environment variable supports feature flags to suppress the
new SQL checks facility. To skip all checks, set to 'true':

  CFDE_SKIP_SQL_CHECKS=true

to skip specific checks by name, set to a JSON dictionary of
check basenames with a truth value:

  CFDE_SKIP_SQL_CHECKS='{"file_check_checksums": true, ...}'

Note, this is parsed as JSON but expects either 'true' or a
dictionary.  Other values may cause errors. To run all checks,
either unset the environment or pass an empty dictionary:

  CFDE_SKIP_SQL_CHECKS='{}'