add registry fixup-fqdn CLI subcommand
The registry contains https URLs to other CFDE resources and these
contain FQDN of the app service.

This tool is for the unusual ops scenario where a registry has been
prepared under an alternate FQDN and then swapped into place by
changes to an ELB config.  This rewrites the various datapackage
and release URLs to point to the new FQDN currently in effect.