fix bug in core_fact.is_bundle ETL, make schema more strict
In the portal, we never want a c2m2 entity row to have NULL core_fact
foreign key, so enforce that and make the build fail if it is
violated.

This was masking a problem with generation of is_bundle because
it wasn't properly included in the core_fact compound key, and so
separate combinations could not be introduced for both
is_bundle=true and is_bundle=false.