update registry with latest vocab term redefinitions
Previously, the registry remembers the first encounter with a new term
used by any submission.  If this definition is less complete, it forever
pollutes the registry UX and we also never see evidence of changes.

Instead, we can favor the most recent usage to prioritize review of
recent submission content.  (Of course, multiple submissions may differ
in definitions and so cause some churn/race conditions this way.)

Adds kludge to only send updates when we detect a difference in the
definition (via client-side comparisons), so that we do not churn the
registry's table histories more than needed during typical ingest
processing, when the same term definitions are encountered repeatedly.