workaround for truncated diagnostic messages in submission listings
The "card" template combines the ingest diagnostic message into a
larger markdown document, but the markdown renderer interprets a
trailing { ... } block as attributes for the resulting HTML rather
than literal text.

By adding an innocuous `.` trailing character in the template, we
ensure that any embedded { ... } will not appear as the final element
of the cell.  A better option would be to invoke a new helper to
apply markdown-escaping to the expanded value in this case.