tighten row-level select policy on ERMrest_Client table
It previously revealed basic display_name and full_name info for users
who are submitting_user of ANY datapackage in the system.  Instead,
only reveal submitting_user info if the client has an appropriate role
with that corresponding DCC.

The access by CFDE-CC are handled by static ACLs for the CFDE-CC
groups, so this dynamic policy is only for DCC users.