remove admin privileges from personal_collection to improve UX
until we can tell Chaise to filter/exclude others' collections from
the chooser modal, we need ermrest to filter it via a more restrictive
row-level policy.

once we can add a domain-filter on the association, we can restore
this select access for admins to be able to review/manage this state.
in interim, only the infrastructure ops (catalog owner) role can
still see other's collections.