enhance model check to tolerate absent nullok=false column w/ default
If a datapackage skips a nullok=false column in the base model and
produces rows with null content in this column, it cannot satisfy the
base model. But, this is not an issue if the base definition includes
a default value, so new rows will get a non-null value populated.