wip: add a custom validator for collection.name uniqueness only run in submission ingests
This requirement for "unique within submission, but not across DCCs"
cannot be expressed as a SQL constraint.  This test, only run in a
single-submission context, is a simple hack to enforce it while
preparing a submission review catalog.

A much more complex query would be to trace through
collection_defined_by_project associations back to the dcc table, so
we could count occurrences per DCC. But, this would still be imperfect
because we do not currently enforce that every collection must have at
least one "defined by" project attribution and therefore do not guarantee
that every collection can be traced to a DCC.