intvert the relationship between CFDE/raw schemas in portal
Create fkeys from CFDE schema tables to the c2m2 schema, in case
we want to project raw values in the UI later:

- CFDE.collection.nid -> c2m2.collection.nid
- CFDE.file.nid -> c2m2.file.nid
- CFDE.biosample.nid -> c2m2.biosample.nid
- CFDE.subject.nid -> c2m2.subject.nid

this patter cannot be followed for _all_ tables because the c2m2
raw submission records are a subset of the CFDE portal records
for some tables, e.g. vocabularies extended by slim maps.