wip: continue revising portal use of the new combined_fact structure
search uses combined_fact.kw, so we do not need to store the extra
keyword data on the other fact tables.

set the frictionless primary keys on the core entity tables to (nid,
combined_fact) to get an index complementary to our (combined_fact,
nid) index requested on each combined_fact fkey column.  the nid
column is still marked required+unique, so the real nid key is still
there as far as deriva is concerned.