adjust ACLs on portal CV tables to allow resource_markdown updates
The columns should be updatable by admin/curator roles running the CLI
tool, requiring that the tables also be updatable. The update right
is suppressed on other columns of the tables and the table also marked
as "immutable" as a hint to Chaise not to show edit controls even for
admins/curators.