tweak text search keyword logic, apply consistently to synonyms
Some CVs have multiword or other compound strings in their synonyms,
so run the same split/trim/merge logic on synonyms arrays that we
already apply to names, descriptions, ids, etc.  Also remove single
character "words" and punctuation which are not going to be useful
with the Chaise searchbox anyway.

This will reduce/simplify the search corpus a bit and hopefully
help a little bit with storage/index efficiency.