adjust text indexing for main entity strings to avoid word-splitting
This is mainly local_id and persistent_id plus a few others that
only exist on file or collection. Avoid word-splitting so that even
long substrings with punctuation will match, i.e. a user pasting
an accession ID or similar.

With the revised keyword table, this change does not have much
impact on the table or index sizes. These per-entity strings are
in a separate keyword record from the vocab-derived keywords...