add phenotypes_flat and diseases_flat dimensions to core_fact
The phenotypes and diseases dimensions store nested structure
tracking [ [vocab_term, assoc_type], ... ] pairs.

The flattened list [ vocab_term, ... ] discards the association
type so that we can use gin_array indexing on the vocab terms. These
could be useful for the kinds of search and aggergation we currently
do in the portal, because we already ignore the association type
in those UX modes.