add 'protein' dimension to StatsQuery2 for completeness
note, this will not work until the prior patch to protein_fact.proteins
is deployed to use an integer array column type.

also, the protein dimension is not currently very interesting as it
just generates a very large number of proteins with one C2M2 collection
for each one, which will not produce an interesting bar graph.