add StatsQuery2.fetch_flattened() method and tests
While not trying to be directly compatible with StatsQuery.fetch,
this converts the multi-termed groups to single terms to support
stats consumers with that limited perspective. Unlike StatsQuery
(based on level1_stats), this StatsQuery2 mechanism uses the fact
tables and only expands groups for the dimension(s) actually
requested by the caller, reducing the multiplicative effects of
"double-counting" compared to level1_stats which pre-expand all
dimensions.