refactor safety checks prior to extractall on archives
Reject unsafe tar archives as an InvalidDatapackage.

- only allow file+dir member types (no links, special devices...)
- check that extraction paths will remain inside the unpack dir

The path check depends on the fact that we extract to an
empty unpack dir and will never create hard or symbolic links
in that unpack dir during extraction.

Out of an abundance of caution, do a similar path check for the
zip archive code path. This seems redundant, if zipfile extraction
does the sanitization described in the docs, where it strips any
absolute path prefix and any '..' parent dir traversals.

Currently, zipfile only supports file+dir members so the other
threat of extracting links is not present. If zipfile gets
such capability, we would want to add a similar pre-check
as in the tar code path, to only allow dir+file members.