bug workaround for 409 conflict when posting submitting user to ERMrest_Client
We ask for onconflict=skip but an ermrest bug causes this to return
a 409 Conflict anyway.  This can happen if the user's name information
is different during cfde-submit than it was when they approached the
portal (for the same globus authentication client ID).

As a workaround, catch the error and double check that the user's
record exists. If so, we can continue anyway... if not, a different
exception will be raised and we will have an ops error.