bugfix for flawed validation check to ensure every project is under DCC hierarchy
The previous query was joining to the project_root table by the wrong
columns. This broken check would accidentally work if the DCC project
was the first row in project.tsv

The revised query avoids using the project_root table at all, in case
we want to deprecate it in the future. It also uses more self-explanatory
set logic (all projects except projects linked to dcc).