change C2M2 file.filename field to required (not null)
As discussed in CFDE-CC, make C2M2 formally require a filename for every
file record. The previous schema specified it as optional, but it is
a de facto required field due to an existing processing bug in the portal
ingest process.

All current DCC submissions do populate the value, so there is little
impact from this change other than improving the error message in the
case that a submission omits filenames.