add portal_prep vocab TSV file paths to registry model and add a protein.tsv.gz
Where we have a built-in, canonical term list, allow registry reprovisioning
to upload those terms to the registry. This will help with subsequent
attempts to upload resource markdown for these term sets.

The protein term list from the ontology working group is far too large
to embed in cfde-deriva (over 3GB compressed with 246M terms). As a
workaround, a smaller extract is embedded which only covers term IDs
that are present in a recent IDG resource markdown file.